implement submitting a answer and it's listing feature

+++ src/main/java/net/slipp/web/QuestionController.java
@@ -48,21 +48,44 @@ public class QuestionController {
 	}
 	
 	@GetMapping("/{id}/form")
-	public String updateForm(@PathVariable Long id, Model model) {
-		model.addAttribute("question", questionRepository.getOne(id));
+	public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.getOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+		model.addAttribute("question", question);
 		return "/qna/updateForm";
 	}
 	
 	@PutMapping("/{id}")
-	public String update(@PathVariable Long id, String title, String content) {
+	public String update(@PathVariable Long id, String title, String content, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
 		Question question = questionRepository.getOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
 		question.update(title, content);
 		questionRepository.save(question);
 		return String.format("redirect:/questions/%d", id);
 	}
 	
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long id) {
+	public String delete(@PathVariable Long id, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.getOne(id);
+		if (!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
 		questionRepository.deleteById(id);
 		return "redirect:/";
 	}
